graph, store: Refile transaction receipt logic to the store crate
